brain/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
comm==0.1.4
debugpy==1.7.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.4.2
easygui==0.98.3
email-validator==2.1.0.post1
etlalchemy==1.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.65.2
fastjsonschema==2.18.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fqdn==1.5.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
install==1.3.5
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.0.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
macholib==1.16.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
moviepy==1.0.3
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
nodeenv==1.8.0
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.26.3
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pbr==1.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
prisma==0.11.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.4.31
pycparser==2.21
pydantic==1.10.13
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
snowflake-id==0.0.5
soupsieve==2.5
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.9.7
SQLAlchemy-Utils==0.30.9
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.14.2
Tempita==0.5.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.13.3
virtualenv==20.25.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
widgetsnbextension==4.0.8
yarl==1.9.4
